+++ CHANGELOG.md
@@ -27,7 +27,7 @@
     * Adding menu option - Adding favorite locations based on location entered - Done!
     * Adding menu option - Adding favorite locations manually - Done!
     * Adding menu option - Editing favorite locations - Done!
-    * Adding menu option - Removing favorite locations - 0% done
+    * Adding menu option - Removing favorite locations - 30% done
     * Adding extra logging and more colors - 0% done
     * Adding new config options to configsetup/setup, and configupdate. - 0% done
 * Adds the ability to manually define a Google Maps API key (because rate limiting has pissed me off too many times) - 0% done